style(PT-49): change styles elements/disabled arrow if carousel refactor(PT-49): Changing folder structure for data feat(pt-49): create data folder
